+++ sdk/monitor/azure-monitor-query/azure/monitor/query/_metrics_query_client.py
@@ -72,8 +72,8 @@ class MetricsQueryClient(object):
          or tuple[~datetime.datetime, ~datetime.datetime]
         :keyword granularity: The granularity (i.e. timegrain) of the query.
         :paramtype granularity: ~datetime.timedelta
-        :keyword aggregations: The list of aggregation types to retrieve. Use `azure.monitor.query.AggregationType`
-         enum to get each aggregation type.
+        :keyword aggregations: The list of aggregation types to retrieve. Use
+         `azure.monitor.query.MetricAggregationType` enum to get each aggregation type.
         :paramtype aggregations: list[str]
         :keyword max_results: The maximum number of records to retrieve.
          Valid only if $filter is specified.
@@ -134,7 +134,7 @@ class MetricsQueryClient(object):
         :keyword start_time: The ISO 8601 conform Date start time from which to query for metric
          namespaces.
         :paramtype start_time: str
-        :return: An iterator like instance of either MetricNamespaceCollection or the result of cls(response)
+        :return: An iterator like instance of either MetricNamespace or the result of cls(response)
         :rtype: ~azure.core.paging.ItemPaged[~azure.monitor.query.MetricNamespace]
         :raises: ~azure.core.exceptions.HttpResponseError
         """