+++ sdk/monitor/azure-monitor-query/tests/test_metrics_client.py
@@ -2,7 +2,7 @@ import pytest
 import os
 from datetime import datetime, timedelta
 from azure.identity import ClientSecretCredential
-from azure.monitor.query import MetricsQueryClient, AggregationType
+from azure.monitor.query import MetricsQueryClient, MetricAggregationType
 
 def _credential():
     credential  = ClientSecretCredential(
@@ -20,7 +20,7 @@ def test_metrics_auth():
         os.environ['METRICS_RESOURCE_URI'],
         metric_names=["MatchedEventCount"],
         timespan=timedelta(days=1),
-        aggregations=[AggregationType.COUNT]
+        aggregations=[MetricAggregationType.COUNT]
         )
     assert response
     assert response.metrics
@@ -34,7 +34,7 @@ def test_metrics_granularity():
         metric_names=["MatchedEventCount"],
         timespan=timedelta(days=1),
         granularity=timedelta(minutes=5),
-        aggregations=[AggregationType.COUNT]
+        aggregations=[MetricAggregationType.COUNT]
         )
     assert response
     assert response.granularity == timedelta(minutes=5)